account page tweaks
