--- color-grading-app/src/pages/SignupPage 2.css
@@ -1,65 +1,0 @@
-.auth-container {
-  width: 100%;
-  max-width: 400px;
-  padding: 2rem;
-  background-color: #1f1f1f;
-  border-radius: 12px;
-  box-shadow: 0 4px 15px rgba(0, 0, 0, 0.3);
-  text-align: center;
-  color: white;
-}
-  
-  .auth-container h2 {
-    font-size: 1.8rem;
-    margin-bottom: 1.5rem;
-  }
-  
-  .auth-form {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-  }
-  
-  .auth-form input {
-    padding: 0.75rem;
-    border-radius: 6px;
-    border: 1px solid #444;
-    background-color: #2b2b2b;
-    color: white;
-    font-size: 1rem;
-  }
-  
-  .auth-form input::placeholder {
-    color: #888;
-  }
-  
-  .auth-form button {
-    padding: 0.75rem;
-    background-color: #4caf50;
-    color: white;
-    font-weight: bold;
-    border: none;
-    border-radius: 6px;
-    font-size: 1rem;
-    cursor: pointer;
-    transition: background-color 0.2s ease;
-  }
-  
-  .auth-form button:hover {
-    background-color: #45a049;
-  }
-  
-  .auth-error {
-    color: #ff4d4f;
-    font-size: 0.9rem;
-    margin-top: 0.5rem;
-  }
-  
-  .signup-page {
-    min-width: 100vw;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    min-height: 100vh;
-    background-color: #121212; /* Optional dark background */
-  }
