--- color-grading-app/src/pages/SignupPage 2.tsx
@@ -1,72 +1,0 @@
-import React, { useState } from 'react';
-import { createUserWithEmailAndPassword, updateProfile } from 'firebase/auth';
-import { auth } from '../firebase';
-import { db } from '../firebase'; // We'll set this up
-import { doc, setDoc } from 'firebase/firestore';
-import { useNavigate } from 'react-router-dom';
-import './Signup.css';
-
-const SignupPage: React.FC = () => {
-  const navigate = useNavigate();
-  const [username, setUsername] = useState('');
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [error, setError] = useState('');
-
-  const handleSignup = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setError('');
-
-    try {
-      const result = await createUserWithEmailAndPassword(auth, email, password);
-      await updateProfile(result.user, { displayName: username });
-
-      // Save user profile to Firestore
-      await setDoc(doc(db, 'users', result.user.uid), {
-        uid: result.user.uid,
-        email,
-        username,
-        createdAt: new Date().toISOString(),
-      });
-
-      navigate('/');
-    } catch (err: any) {
-      setError(err.message || 'Signup failed');
-    }
-  };
-
-  return (
-    <div className="signup-page">
-      <div className="auth-container">
-        <h2>Create an Account</h2>
-        <form onSubmit={handleSignup} className="auth-form">
-          <input
-            type="text"
-            placeholder="Username"
-            value={username}
-            required
-            onChange={(e) => setUsername(e.target.value)}
-          />
-          <input
-            type="email"
-            placeholder="Email"
-            value={email}
-            required
-            onChange={(e) => setEmail(e.target.value)}
-          />
-          <input
-            type="password"
-            placeholder="Password (min 6 characters)"
-            value={password}
-            required
-            onChange={(e) => setPassword(e.target.value)}
-          />
-          <button type="submit">Sign Up</button>
-          {error && <p className="auth-error">{error}</p>}
-        </form>
-      </div>
-    </div>
-  );
-};
-
-export default SignupPage;
