+++ color-grading-app/src/pages/SignupPage.tsx
@@ -1,11 +1,15 @@
 import React, { useState } from 'react';
-import { createUserWithEmailAndPassword } from 'firebase/auth';
+import { createUserWithEmailAndPassword, updateProfile } from 'firebase/auth';
 import { auth } from '../firebase';
+import { db } from '../firebase'; // We'll set this up
+import { doc, setDoc } from 'firebase/firestore';
 import { useNavigate } from 'react-router-dom';
-import './SignupPage.css'; // Ensure you have a CSS file for styling
+import './SignupPage.css';
+
 
 const SignupPage: React.FC = () => {
   const navigate = useNavigate();
+  const [username, setUsername] = useState('');
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
   const [error, setError] = useState('');
@@ -15,8 +19,18 @@ const SignupPage: React.FC = () => {
     setError('');
 
     try {
-      await createUserWithEmailAndPassword(auth, email, password);
-      navigate('/'); // Redirect to home (or dashboard/editor)
+      const result = await createUserWithEmailAndPassword(auth, email, password);
+      await updateProfile(result.user, { displayName: username });
+
+      // Save user profile to Firestore
+      await setDoc(doc(db, 'users', result.user.uid), {
+        uid: result.user.uid,
+        email,
+        username,
+        createdAt: new Date().toISOString(),
+      });
+
+      navigate('/');
     } catch (err: any) {
       setError(err.message || 'Signup failed');
     }
@@ -24,27 +38,34 @@ const SignupPage: React.FC = () => {
 
   return (
     <div className="signup-page">
-    <div className="auth-container">
-      <h2>Create an Account</h2>
-      <form onSubmit={handleSignup} className="auth-form">
-        <input
-          type="email"
-          placeholder="Email"
-          value={email}
-          required
-          onChange={(e) => setEmail(e.target.value)}
-        />
-        <input
-          type="password"
-          placeholder="Password (min 6 characters)"
-          value={password}
-          required
-          onChange={(e) => setPassword(e.target.value)}
-        />
-        <button type="submit">Sign Up</button>
-        {error && <p className="auth-error">{error}</p>}
-      </form>
-    </div>
+      <div className="auth-container">
+        <h2>Create an Account</h2>
+        <form onSubmit={handleSignup} className="auth-form">
+          <input
+            type="text"
+            placeholder="Username"
+            value={username}
+            required
+            onChange={(e) => setUsername(e.target.value)}
+          />
+          <input
+            type="email"
+            placeholder="Email"
+            value={email}
+            required
+            onChange={(e) => setEmail(e.target.value)}
+          />
+          <input
+            type="password"
+            placeholder="Password (min 6 characters)"
+            value={password}
+            required
+            onChange={(e) => setPassword(e.target.value)}
+          />
+          <button type="submit">Sign Up</button>
+          {error && <p className="auth-error">{error}</p>}
+        </form>
+      </div>
     </div>
   );
 };
